replace component with hook
